+++ src/components/Field/FieldTypes.tsx
@@ -11,7 +11,7 @@ import { FieldDefMapCoordinates } from "@root/components/Field/FormFieldMapCoord
 import { FieldDefMatrix } from "@root/components/Field/FormFieldMatrix";
 import { FieldDefNumberTable } from "@root/components/Field/FormFieldNumberTable";
 import { FieldDefPhoneSelection } from "@root/components/Field/FormFieldPhoneSelectionDropdown";
-import { FieldDefRadio } from "@root/forms/FormFieldRadio";
+import { FieldDefRadio } from "@root/components/Field/FormFieldRadio";
 import { FieldDefRaw } from "@root/forms/FormFieldRaw";
 import { FieldDefTable } from "@root/forms/FormFieldTable";
 import { FieldDefText } from "@root/forms/FormFieldText";
